+++ myrocketsimulator/data/defaultMRSmission.py
@@ -62,9 +62,9 @@ class MRSmissionData():
     # All values are defined in ICRF/J2000 and are [m]/[m/s].
     # The default mission provides a state vector of the ISS for UTC provided above.
     y0 = np.array([
-        6091.160573649400, 2321.823227910050, -1926.816516037270 ,   
-        -0.21429216175361, 5.21216680526399, 5.60755423362593        
-        ]) * 1000 # multiply by 1000 to get meters / meters per second
+       6091.160573649400, 2321.823227910050, -1926.816516037270 ,   
+       -0.21429216175361, 5.21216680526399, 5.60755423362593        
+       ]) * 1000 # multiply by 1000 to get meters / meters per second
     
     # t0_MET defines the MET-value at the moment of t0_JD/t0_UTC. Default is 0.
     # The use depends from the launchtype: